requirements.txt
streamlit
google-genai
langchain
langchain-community
langchain-google-genai
chromadb
pypdf